Updated branche to stay aligned with the latest master version.

+++ scooter-service/src/main/java/ar/edu/unicen/scooterservice/domain/dtos/response/StopResponseDTO.java
@@ -4,8 +4,8 @@ import ar.edu.unicen.scooterservice.domain.entities.Stop;
 
 public record StopResponseDTO(
         Long stopId,
-        int latitude,
-        int longitude
+        float latitude,
+        float longitude
 ) {
     public static StopResponseDTO toDTO(Stop stop) {
         return new StopResponseDTO(
